webroot/dqb/dqb/requirements.txt
conf==0.4.1
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
mysqlclient==1.4.4
PyMySQL==0.9.3
pytz==2019.2
sqlparse==0.3.0
